requirements.txt
evaluate==0.4.0
info-nce-pytorch==0.1.4
nltk==3.8.1
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvcc-cu11==11.8.89
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
pandas==1.5.3
rouge-score==0.1.2
sacrebleu==2.3.1
scikit-learn==1.2.1
scipy==1.10.0
sentence-transformers==2.2.2
tqdm==4.64.1
transformers==4.28.1
wandb==0.15.0